Notate sqrt(2) proof theorems with suggestions for important changes
Notate sqrt(2) proof theorems with suggestions for important changes, such as where the theorems should be moved (i.e. in which context each should appear), how to modify them in many cases to make them more general, and how to modify or augment related automation to take advantage of the theorems. Such changes are now part of the "to do" list. Related Issue: #93.