Change dividesInBool axiom to a theorem and update related files
Move the dividesInBool axiom from the number/divisibility context axioms to theorem page and update the Divides.deduceInBool() method to accommodate the change. A next step here will be to provide an axiomatic definition for “a|b” (a divides b) in the axioms, but at this time the axioms remain empty. Related Issue: #93, #179.