Correct misspelled method name in several related files
The deriveSupersetMembership() method appears in a variety of places (mainly in logic/set_theory/containment classes) but variously spelled sometimes as deriveSupsersetMembershipCorrect (with an extra 's') and sometimes correctly as deriveSupersetMembership. This commit corrects all the extant misspelled versions, including places where the misspelled version was being called. Related Issues: #93.