Add deduce_member_non-zero method for naturals side-effect
Related Issue: #93. Added a deduce_member_non-zero() method in NaturalsPos class to be called from the from the membershipSideEffects() method. Not absolutely clear this is necessary or desirable, but at least temporarily addressing a general issue in developing the irrationality of sqrt(2) proof where positive nat is not automatically known to be non-zero.