Establish GCD_one_def thm in number/divisibility module
Establish GCD_one_def thm in number/divisibility module. This thm concludes from a GCD(a,b)=1 that a and b are relatively prime (giving the result more generally in terms of positive natural p > 1 such that Not(p|a AND p|b)). No proof of the thm established yet. Related Issue: #93.