Sqrt(2) irrational proof fixes.
"divides_if_divided_power" was an incorrect theorem (as Bob Carr pointed out).  For the purpose of the sqrt(2) proof, it was replaced with "even__if__power_is_even".  Also the sqrt(2) irrational theorem needed to be moved later in the exponentiation theorems page because it depends upon other exponentiation theorems.