Implies.conclude fix: reintroduce 'evaluations'
Calling 'evaluated' (or 'evaluation') is designed to be not so time-consuming.  It is trivial compared with 'simplification' because it gives up if it is unable to evaluate anything.  It is proper to try evaluating the antecedent and consequent when they aren't already proven/disproven.  (An older version of Implies.conclude used it.)