Divides automation fix/modification
In a side-effect test,
Equals(self.lhs.exponent, self.rhs.exponent).proven()
needed to be changed to
self.lhs.exponent == self.rhs.exponent
to really meet the requirement (not just logical equality, but it needs to be the same expression).

In Divides.conclude, remove "case (5): x|y if x^n|y^n".  This should just be handled on the side-effect side.