Existential theory package updates: axiom mods and skolemElim proof
Added a proof of skolemElim.  In the process, issues with the existential axioms became evident.

Modifying existential axioms to generally include conditions seemed to be necessary, at least for existsDef; otherwise, you end up with thinks like
\exists_{x | Q(x)} P(x) = \lnot \forall_x {P(x) if Q(x). \neq True
and
{P(x) if Q(x). != T
is deliberately left undefined when Q(x) is untrue.  However, since
[\exists_{x | Q(x)} P(x)] = \lnot [\forall_{x | Q(x)} [P(x) \neq T]]
it follows that
[\exists_{x} P(x)] = \lnot [\forall_{x} [P(x) \neq T]]
but then we can instantiate this while adding back a conditional such that
[\exists_{x | Q(x)} P(x)] = [\forall_x {P(x) if Q(x). != True]
  = [\forall_{x | Q(x)} {P(x) \neq True]

So, {P(x) if Q(x). != T may not be intrinsically defined  on its own, it does have a reduction within the context of universal quantification.  We need to be careful about consistency in this regard.