Add warning comment for exp_neq thm in number_exponentiation
The exp_neq thm in number_exponentiation appears to be problematic, claiming that for all a, x, y in reals, Exp(x, a) ≠ Exp(y, a) when x≠y, but clearly when we have things like x = -1, y = 1 and a = Even integer, we have x^a = y^a. Related Issues: #93.