Preparation for allowing ordering reversal as style invariant.
Updated the Relation base class of TransitiveRelation.  The "lhs" and "rhs" attributes are no longer real dictionary attributes but obtained on-the-fly via "__getattr__" and now depends upon the "direction" style ('normal' or 'reversed').  Updated LesserRelation.lower() and LesserRelation.upper() to be correct even if the style is "reversed".  This is just some preliminary preparation for making ">" be a style variant of "<" and "superset" be a style variant of "subset".