Context storage overhaul / Prove-It database reorganization.
Added ContextFolderStorage class.  Reference counts have been eliminated in the Prove-It database.  Instead, there is a separate folder generated by distinct notebooks -- when "%end" or "%qed" is called, the folder is cleared of everything that is not referenced in memory.  Axiom and theorem references have been changed to the KnownTruth rather than Expression (this was mainly motivated by the "trueAxiom" axiom so that it would be distinct from the "TRUE" common expression).