Create "canonical" version of expressions only on demand.
The canonical ("generic" is the old term that is now obsolete) version of an expression is only needed when we need its hash value (checking/adding the expression in a python set or as a dictionary key) or we are comparing expressions via __eq__.  Don't use expressions in a manner that requires the hash when it can be avoided: changed _StyleData.parent_map structure to avoid this, and only check if Literals are in Operation.operationClassOfOperator.  Also, disable automation when creating the canonical version of a lambda expression.