Fixed theorem renaming feature as well as alternate proof feature.
Other changes had broken the recently added "theorem renaming"
feature.  Also fixed it to properly handle the case when you
have multiple names for the same theorem expression for an
alternate proof, as well as renaming one or the other in this
case (it goes by the order in which they appear).