Moved theory notebook files into '_theory_nbs_' folders.
This as motivated by a name collision with 'numbers'
(our theory package and something used by Jupyter causing notebooks
at the 'proveit' package level to no longer run).  By having the
notebooks in sub-folders we avoid this issue.  It is also better for
organization and alleviates the need for the underscores around
_common_, _axioms_, _theorems_, and _demonstrations_.  These
underscores served the purpose of keeping these special notebook
files together in a sorted list, but having just a few files in
a sub-folder is even better.