'python build.py --all' require demonstration and theorem proof re-execution.
The problem isn't fully understood, but for now we solve it be execting
the demonstration and theorems proofs again after re-execution of theorem
notebooks.  There may be some amount of randomness in the theorem notebook
execution that may invalidate demonstration and theorem proofs so their
databases are no longer accurate until the are executed again.