Update QPE/axioms.ipynb with an additional Circuit
This extends the partial definition of the QPE circuit definition begun previously with the help from JML, now starting the definition of the recursive portion of the QPE circuit definition. This is not yet correct, but it does compile and does show a Circuit that gives the right idea. We need a little extra machinery in the Circuit class to get much better.