Add Ceil(x) is Real, Floor(x) is Real, Round(x) is Real thms to rounding
Add Ceil(x) is Real, Floor(x) is Real, Round(x) is Real thms to rounding package and prove the theorem for the Floor case: the Floor of a real value is itself real. The proof should eventually be revisited to make it shorter -- currently the proof involves an indirection using Floor is Z to Z is Q to Q is Real but should be able to go directly from Z to Real. Related Issue: #97.