Upgrade side-effects for IntegerSet to include inclusion in Real
Upgrade side-effects for IntegerSet to include inclusion in Real: by adding a deduce_membership_in_real() method and adding a call to that method in the membership_side_effects() method. This should shorten a number of simple proofs where with an integer x but want to know that x is also a real number. Previous to this, proofs were having to go from Integer to Rational and then from Rational to Real and this should allow us to go directly from Integer to Real. Related (indirectly) Issue: #97.