Modify the X_of_real_is_real thms for X = ceil, floor, round
Modify the X_of_real_is_real thms for X = ceil, floor, round to revert back to not using the automatic Z-to-Real deduction and instead using the automatic chaining Z-to-Q-to-Real. Will create an Issue for the broader question about how aggressively to pursue the membership side-effects, with WW prefering right now to focus on the proofs with the already-established side-effect/automaticity.