Number ordering reversal fixes
There is a problem in that the 'reversed' method does work as one might expect when calling it on a Judgment since it doesn't start as 'with_'.  Perhaps this abbreviation of 'with_direction_reversed' should be eliminated, or perhaps we should make Judgment.__getatt__ a little more sophistocated.  Later though.