Establish proof of pos_domain_in_full_domain_mem thm in QPE package
This commit includes a re-naming of the earlier (and commented-out) membership-based pos_domain_in_full_domain thm to pos_domain_in_full_domain_mem and its eventual proof. Eventually might want to rename this back as the default name and the subset_eq-based version of the theorem renamed or deleted. Related Issue: #97.