Update the pos/neg_domain_in_full_domain thm proofs
Update the pos/neg_domain_in_full_domain thm proofs, temporarily commenting-out portions because the proofs no longer work with recent changes in axiomatic definitions of Interval class and related Interval theorems (in the axioms and theorems in which an Interval(a, b) appears, we must now have a ≤ b). Related Issue: #97, #147.