Establish IntervalMembership class and related thms and testing
This is a first attempt to establish the IntervalMembership class and related automation methods, along with a number of related theorems in the number_sets/integers package. None of the theorems have been proven yet. The package demonstrations page now has a goodly number of cells testing the Interval class and related membership methods and automation. Pre-existing Interval-related theorems have also been updated to require a ≤ b in Interval(a, b) domains, reflecting recent additions to the axiomatic basis for the integral Interval structure. Related Issues: #97, #147.