Add real IntervalXX thms to number_sets/reals package
Add real IntervalXX thms to number_sets/reals package. These generally involve real interval membership properties, such as x not being in a real interval if x is not real or a <= x <= b if x is in the real interval [a, b], etc. Some of these may be utilized in the RealIntervalMembership (and Nonmembership) class. Related Issue: #97, #147.