Update number_sets/integers demo nb for removing a<b constraint
Update number_sets/integers demo nb to acknowledge the removal of the a <= b constraint on integer intervals of the form Interval(a, b). This can create some odd-looking but vacuously true results like (x \in {8..3} |- 8 <= 3), but the lack of transparency in such results seems a good tradeoff for the simplicity of not having to always prove the a <= b constraint when applying related theorems. Related Issue: #97.