Fixed lambda_expr._replaced problem that was recently introduced.
The problem is that, in an effort to be more efficient, I decided to exclude keys from the inner_repl_map that are variables that aren't free in the body.  But sometimes we want to propagate it anyway to relabel variables that aren't free.  That was a simple fix, but I also changed the way that Lambda._replaced deals with making replacements of parameter indices up-front, making it a little easier to understand.