Fix Lambda._replaced again... wasn't quite right
Actually, there was a good reason why we were excluding keys from the
inner_repl_map with variables that aren't free in the body -- we wanted to avoid the appearance of collisions that aren't actual collisions.  So we do exclude them from inner_repl_map, but we include them in relabel_map.  That way, we can still relabel those variables even if they aren't free in the body and also avoid the appearance of a variable collision that isn't real.  A couple of related fixes: proper handling of relabeling IndexedVar's to avoid collision, and properly reducing parameter ranges (e.g., including requirements).