Add general summation-splitting thm to summation theory package
Related Issue: #97. Add general summation-splitting thm to summation theory package, a somewhat more general version of what was already available (which also then depends on the assumption that a sum from index = a to index = b is 0 if b < a). Helpful in recent/current work on alternate proof of Nielsen & Chuang's identity 5.20.