Add simplify options to Sum.shift() method
Related Issue: #97.  Add 2 simplify option args to the Sum.shift() method and underlying machine inside the method to allow for the simplification() to be called on the shifted indices and shifted summands resulting from the index shift operation. This commit also modifies a fence=fence to fence=True formatting argument in the _formatted method to more properly parenthesize things like Sum(x+1) so it doesn't look like Sum x + 1.