Further implement reduction formulas in Sum.shift() method
Related Issues: #97, #224. Tidy up some of the effort to utilize reduction formulas for shifted indices and shifted summand in the Sum.shift() method and send such reduction formulas to the eventual instantiation of the index_shift theorem. Also updates description of the Sum.shift() method and adds a NaturalPos option to the Sum.deduce_in_number_set() method.