Modify Expr.simplification() to take and use a shallow=T/F arg
Related Issues: #97, #224. A possibly temporary modification of the Modify Expr.simplification() method to take and use a shallow=T/F arg, with default setting shallow=False, and which when shallow=True causes the first try block of "default tricks" to be skipped and also causes any resulting known simplification to not be stored for future use. The overall modification of the Expr.simplification() method is driven in part to allow a simplification() call ot be made from within a reduction-producing sub-code in the Sum.shift(), Sum.split(), and Sum.join() methods. Still need to more carefully consider the except catching of the try-generated errors.