A few small exponentiation package updates
Corrections to exp_zero_eq_one, exponentiated_zero, and exp_real_pos_closure theorems (proper conditions/domains).  Update Exp.do_reduced_evaluation to handle the case of a literal positive integer exponent and literal base.  Currently this only works up to an exponent of 9, so that will need to be updated in the future.  A fix in Exp.distribution when the base is a non-binary product.