Core updates: propagate styles through replacements.
When rebuilding an expression with replaced sub-expressions, use the original styles as applicable.  In the process, I created a TemporaryDefaults context manager class for temporarily setting default styles, but this with be useful more generally.  Also, instead of having default styles for expressions with the same meaning, changed this to be default styles for expressions with the same labels and meaning.  Additionally, fixed some things w.r.t. canonical expressions (avoid building a canonical expression of a canonical expression -- it should be final and not change).