Core fixes w.r.t. presumptions and CircularLogic.
No longer have CircularLogic inherit from ProofFailure so that these
exceptions will not be ignored when deriving side-effects.  Raise
this exception whether the infraction is caused by an explicit
or implicit presumption (with the specific theorem presumed or just
its package).  This choice is because reversing a presumption with
recently introduced "efficiencies" wasn't correct and it is easier
just to fail and force the user to add it to the list of
presumption exclusions.