Mandatory 'styles' keyword-only argument for Expression constructors
Rather than using "defaults.styles" as a "backdoor" means of setting initial styles (which turns out to be a dangerous approach because it can be applied to other expressions unintentionally), your a "styles" keyword argument for all __init__ methods of class derived from Expression which propagate to Expression.__init__.  An exception is raised if this keyword-only parameter is forgotten.  Also, updated inspect.getfullargspec to inspect.signature (which is deprecated in Python 3.8).
Made some 'Operation' -> 'Function' changes where appropriate.  A couple minor changes pertaining to proveit magic commands (clear 'definitions' when "%begin" executes and use OrderedDict for the 'definitions' instead of a separate ordered list of names).