Update SubIndexed class operator latex_format
Related Issue: #97. Provide a latex_format version for the literal operator of the SubIndexed class. The SubIndexed class provides subscript indexing of a literal (in contrast to the IndexedVar, which provides subscript indexing of a variable). Without a latex_format specified for the _operator_, the occasional listing of the SUB_INDEXED in a DAG printout was interpreting the underscore as a subscript command; supplying the appropriate latex_format version seems to now fix this. Also update the description comments. Would probably be good to eventually move this class elsewhere for more general usage.