Defining @prover and @equivalence_provers.
A @prover method will handle changing defaults temporarily via keyword arguments and checking that the result is a Judgment.  An @equivalence_prover does these and additionally "registers" it for use with InnerExpr, checks that the result proves an equality with self on the left side, and performs handles some basics for 'simplfication' and 'evaluation' @equivalence_provers (remembering and recording simplfications/evaluations and checking that 'evaulation' right sides are irreducible values).