Update Mult class w/misc fixes (see comments)
Related Issues: #97, #228. Update/augment Mult class methods in several ways: (1) correcting a repeated typo where the cancelation() method was spelled as cancellation; (2) augment Mult.cancelation() method so it removes a final factor of 1 produced by a cancelation (producing |- old = x instead of |- old = 1*x) and acknowledges the possibility of the final rhs result being a single entity that is not a Mult; and (3) registering the Mult.cancelation() method as an InnerExpr equivalence method.