@equivalence_prover -> @equality_prover
General substitution is only applicable for equality, not general equivalence.  Having a "present tense" form of a method that performs substitution of an inner expression within a Judgment to produce a new Judgment is only applicable to equality -- so equality is special.  Next I'll add @relation_prover for more general purposes.