Add add_one_left_in_exp_poss_zero_base thm to exponentiation package
Related Issue: #97. Augment the numbers/exponentiation package with the add_one_left_in_exp_poss_zero_base theorem, allowing a(a^b) to go to a^{b+1} even when a might be zero (but then constraining b and b+1 to be non-zero to avoid 0^0 controversy). Still also have the add_one_left_in_exp theorem allowing a(a^b) to go to a^{b+1} when a != 0. The new theorem is then utilized in the proof-by-induction of the finite_geom_sum theorem.