Redefine psi_t to utilize variable t instead of literal t
Related Issue: #97. Redefine psi_t to utilize variable t instead of literal t. Although t represents the number of qbits and could be considered as a given constant, it is more useful to consider t as a variable in NaturalPos, which also then allows the appropriate expression of various theorems involving psi_t.