Update inclusion/demos nb, up to testing of ProperSubset.apply_transitivity()
Related Issue: #225. Update sets/inclusion/demos nb, up to section testing the ProperSubset.apply_transitivity() method, which currently still prompts an error and will need to be addressed in the .py file itself. This includes updating of a number of cells previously commented out and some re-working of some of the demonstrations/examples.