Reverting recent changes to inner_expr.py
Using "**defaults_config instead of assumptions keyword in 'inner_equiv' method in inner_expr.py" was not really a good idea. This breaks things when there are other keyword arguments besides configurations of the defaults. It is find how it was -- only 'assumptions' need to be pulled out when doing a 'substitution' because substitutions use preserve_all=True anyways.