Update EnumMembership.conclude() w/auto_simplify=False flag.
Related Issue: #225. Update EnumMembership.conclude() w/auto_simplify=False flag when instantiating the in_enumerated_set theorem. It's not quite clear why this needs to occur, but it does eliminate some errors that can arise in the InSet.conclude() method when dealing with sets containing elements that are not irreducible.