Split core-level equality replacements into two passes
In the first pass, we make manual replacements. In the second pass, we make auto-simplification replacements. By splitting this into two passes, expressions can be "preserved" more properly -- sub-expressions may be replaced to get it into a preserved form so it won't be auto-simplified in the next pass. Note: it is no longer the case that a replaced expression won't ever be auto-simplified; now, with these two separate passes, it is possible to replace an expression and then auto-simplify the replacement. The replacement will have to be explicitly preserved to prevent such a thing. Of course, if manual replacements are used for the purpose of converting an expression to a conserved form, this is not a problem.