Correct the set of count_to_x_range thms in numbers/numerals/decimals
Related Issue: #225. Correct the set of count_to_x_range thms in numbers/numerals/decimals so that (for example) the count_to_4_range theorem reads as (1, 2, 3, 4) = (1,...,4) instead of (2, 3, 4)=(1,...,4). The latter (incorrect) version was introduced inadvertently when trying to eliminate the count_to_1_range involving a an ExprRange with a single element.