Added 'deduce_in_number_set' function
which attempts to use the 'deduce_in_number_set' method if one exists; otherwise, it builds the proper InSet expression and calls 'prove'. This is useful to call when automation is turned off in order to nudge things a little in the right direction and allow them to go through anyways.