Don't replace and preserve the same expression.
Fixed Operation.simplification and Operation.evaluation to not
make a replacement out of the reduction when the reduction is
trivial in order to avoid making a replacement of a preserved
expression (the original expression).  Added checks in
decorators.py to make sure an expression is not simultaneously
replaced and preserved.