New InnerExpr features to automatically handle instance substitutions...
when making substitutions inside Lambda expressions when the Lambda parameters are involved.  The conditions of containing Conditionals are automatically assumed when proving an inner equality.  Also, when making an InnerExpr via Judgment.inner_expr(), the Judgement assumptions will automatically get used.  Made InnerExpr.substitution and InnerExpr.substitute be @provers, and made it so its 'assumptions' (e.g. from Judgement.inner_expr()) of an InnerExpr are accounted for in decorators.py.