Establish psi_prime_t_lit_formula theorem and proof
Related Issues: #97. Establish and prove the psi_prime_t_lit_formula theorem. This is the literal t version of the psi_prime_t_formula thm, which itself represented Nielsen & Chuang's formula 5.20 (pg 222). The literal t (accessed in the notebook from the QPE common nb as t_) represents the size (in number of qubits) of the first register.