Establish and prove scaled_abs_delta_diff_interval_alt QPE thm
Related Issues: #97, #225. Establish and prove scaled_abs_delta_diff_interval_alt QPE thm. This is an alternative for the scaled_abs_delta_diff_interval in which we now assume that ell ≠ 0 (which allows us to use some previously-proven results and in effect prove the domain to be the half-open interval (0, pi/2] instead of the closed interval [0, pi/2]. Still needs some cleaning up.