Add ExprTuple.range_expansion() method
Related Issues: #97, #225. Add range_expansion() method to the ExprTuple class, allowing ExprTuple of a single ExprRange representing a small finite number of elements to be equated to an ExprTuple with the finite number of elements appearing explicitly as ExprTuple args. Not yet tied in to any auto_simplification machinery. The new method requires a new set of theorems in the numbers/numerals/decimals package (see separate commit for those).