Enable substitution of operands as ExprTuples
Implementing Issue #236.  We allow a non-ExprTuple (typically a Variable)
to represent the ExprTupe of operands of an Operation via the 'operands'
keyword argument in the Operation/Function constructor.  We can then
substitute the 'operands' of an inner expression as a whole.  Using this
new feature in Equals.conclude_via_direct_substitution and demonstrating it.