Establish Mult.common_power_extraction() method
Related Issues: #97, #225. Establish the Mult.common_power_extraction() method to take a product of exponentials and rewrite as a product raised to some common exponent factor (for example taking a^{ij} b^{jk} to (a^i b^k)^j). This now shows up in the testing in multiplication/demonstrations.ipynb and is used in the proof nb for the alpha_l_summed theorem in the quantum/QPE pkg.