Update Exp.distribution() method to anticipate more possibilities
Related Issues: #97. Update Exp.distribution() method to anticipate more possibilities. In particular, if exponent is a NaturalPos, then we shouldn't necessarily require the instantiation of the posnat_power_of_posnat_power theorem, instead allowing for the possibility that we can use exponent in a superset (such as Natural, Real, etc). This makes for some redundant-looking code that should probably be reorganized later.