Update w.r.t. on-the-fly 'both_sides' methods of Relation
In addition to looking at method names of sets that are known to contain the left and/or right side, also look at the method names of the left/right side expression classes.  In fact, these take priority.  Also, use functools @wraps to ensure the on-the-fly method has the proper name/docstring.