Establish delta_diff_exp_not_1 thm; update domain for alpha_l_sqrd_ineq
Related Issue: #97. Establish delta_diff_exp_not_1 thm (to be used in alpha_l-related theorems); update the domain and conditions for alpha_l_sqrd_ineq thm (to match the use of full domain with ell != 0 utilized in previous theorems); and update the format of the alpha_l_summed_abs thm to include the trigonometric  denominator.