Update QPE theorems page and related theorem nbs
Related Issue: #97. This includes some renaming and deletion of theorems as I try to reorganize and combine a few proofs. Some of this was then short-circuited or partly abandoned when it became clear that I needed to go ahead and merge one the working repos of @wwitzel into my own to update quantum algebra-related stuff before working too much more on QPE theorems.