Fix relabeling of parameter ranges and related enhancements
Fixed a bug in which it was not properly relabeling parameter
ranges and added functionality to allow relabeling of such a
range by indicating just the variable in the repl_map key.
Also raises a ValueError if trying to instantiate/relabel
a variable that is not instantiatable or relabelable in the
Judgment.  Implements #245.