Begin to update _alpha_l_summed thm proof for recent notation changes
Related Issues: #97, #247. Begin to update _alpha_l_summed thm proof for recent notation changes. About half-way through getting a "dictionary keys changed during iteration" error, which we're hoping will be cleared up with a merging of one of WW's branches into mine for a fix. So think of this as a temporary commit.