No partial masking of parameter ranges allowed anymore
For example, given a Lambda expression
   (x_i, ..., x_n) -> f(x_j, ..., x_m)
we effectively assume that i <= j and n >= m so there are no
occurrences of x where the index is ouside of the introduced
range from i to n.  The Lambda introduces variables in a new
scope, masking their external definitions.  When introducing
a range of variables, the masking is assumed to be complete and
this is verified when it matters -- when attempting to instantiate
or relabel ranges using "multiple equivalent expansions".

This simplified the core somewhat and made the 'err_inclusively'
argument of 'free_vars' obsolete.  It also means that simple
variable relabeling of Lambda expressions is always allowed.
For example,
   (x_i, ..., x_n) -> f(x_j, ..., x_m)
is the same as
   (y_i, ..., y_n) -> f(y_j, ..., y_m)
because we can safely assume full masking.