Established alt version of equiv_sums summation thm
Related Issue: #97. Established alt version of equiv_sums summation thm to better specify the need for the sets S and R to be finite and have the same cardinality (implicitly then demanding that the mapping g from S to R is bijective). Left the old equiv_sums theorem just in case and named this new theorem equiv_sums_alt for the time being.