Change fb to f__b_1_to_j in tensor distribution theorem
Related Issues: #97, #245. Correct the functional notation in a specific spot in the tensor_prod_distribution_over_summation theorem in the tensors package: f(b) was inadvertently being used instead of f(b1, ..., bj), and this was causing strange things to happen during instantiations where the f(b) stayed around in the resulting instantiation.