Modify InnerExpr._substitution() to elim unneeded assumptions
Related Issue: #97, #256. Modify InnerExpr._substitution() to eliminate extraneous assumption(s) when doing an "instance" substitution (for example, the range of index values for a summation). Before this change, for example, a substitution for the summand in a summation kept as an assumption the values of the summation index, but that assumption is redundant since the summation itself implicitly proscribes the index values.