Add bounding thms for exponent=2 in exponentiation pkg
Related Issue: #97. Add two bounding theorems in exponentiation package to deal with squaring when we have a negative base: exp_even_neg_base_less and exp_even_neg_base_lesseq, analogous to a set of other theorems already in that package for dealing with exponentiation of a non-negative base. These new theorems are then used to augment the Exp.bound_via_operand_bound() method, which is then further demonstrated/tested in the exponentiation demonstrations nb. An oddity seemingly required in the eventual method code is somewhat worrisome: checking that the exponent == two rather than being able to check that Equals(exponent, two).proven(); I am not sure of the implications of that bit of coding subtlety.