Add conjunction_over_expr_range thm to booleans/conjunction pkg
Related Issue: #97. Add conjunction_over_expr_range thm to the logic/booleans/conjunction package. Intended to then be utilized somewhere in the And.conclude() method (somewhat analogous to the use of the And.conclude_as_redundant() method and the redundant_conjunction theorem), and should help with an error arising in a proof step for the _psi_t_var_formula proof, but not yet implemented.