Add even_fxn_sum thm to numbers/summation pkg
Related Issue: #97. Add even_fxn_sum thm to numbers/summation pkg to allow the negation and reversal of summation indices when summand is a function squared (meant now to also be generalized later when we have defined the class of even functions). This was then utilized to take a step in the proof of the QPE theorem _fail_ineq (see Nielsen & Chaung's eqs (5.30)-(5.34)). This thm is a special case of a more general even_fxn_sum theorem previously utilized in the old proof of the _fail_ineq theorem.