Correction to 'note' about integer Intervals equiv to empty set
Correction to 'note' about integer Intervals equiv to empty set. The original note had the claim backward but we wanted to say that if b < a then the Interval(a, b) is equivalent to the empty set. See the 'vacuous_interval' theorem in the numbers/number_sets/integers pkg theorems.