Establish the numbers/functions subpackage
Related Issue: #97. Establish the numbers/functions subpackage, including a few items in the numbers/function/common nb, all of which is related (or is planned to be related) in the near-term to the proof of the QPE theorem _fail_ineq, which needs some infrastructure for integrals of certain classes of functions (e.g. even functions and monotonically-decreasing functions).