Establish And.conclude_over_expr_range() and augment And.conclude()
Add the And.conclude_over_expr_range() method (which uses the recently-created conjunction_over_expr_range theorem from the logic/booleans/conjunction package) and augment the And.conclude() method with this additional method (paired with an analogous method called conclude_as_redundant). This help solves an error that was arising in one of our QPE proof efforts (_psi_t_var_formula) in proving vector space memberships for elements of an ExprRange.