Number operation “factorization” improvements
Factor multiple factors from a Mult that do not have to be consecutive. This is done in a smart way, with group commutations where it can. Similarly, changed Divide.eliminate_common_factor to Divide.eliminate_common_factors (plural) with functionality to eliminate any number of common factors which do not have to be consecutive.