Using "parts" in quantum circuit multqubit operations
Instead of using ExprTuples as ordered sets for the multi-gate/input/output/measure qubits, use regular set (Intervals in this case specifically), but indicate a part number for each element (part 1, part 2, ...).  That just makes it more clear what is going on.  Also, used NamedExprs for the QcircuitElement operands (which may optionally of a "part" keyword).