Add VecAdd.factorization() method
Add VecAdd.factorization() method, which then deals with two special cases: a VecAdd of either (1) all ScalarMult operands or (2) all TensorProd operands. Generalization of this method to handle other operands (for example, VecSums) and a mixture of different types of vector expressions is currently being developed. The VecAdd.factorization() method is utilized in the proof of the QPE theorem _psi_t_var_formula.