Fix 'with_matching_style(s)' for Expressions and Judgments
For styles to match, just use the expression with the new style to make sure it includes differences of internal labels. All we have to do is check that the expressions have the same meaning (are '=='). Note that "with_matching_style" is automatically called by the @prover decorator if it is a 'conclude' prover, so we removed redundant calls in some of these methods.