Forall side-effect / conclude enhancement w.r.t. generic/canonical forms
In Forall.conclude_via_generalization, check to see if there is a canonical form that is proven which can simply be generalized. On the other side, as a Forall side-effect, do a generic instantiation of the original and canonical form but turn off automation while doing it to avoid cascading side-effects that get out of hand. The generic instantiation goes to all levels of nested universal quantification with the default instantiations (parameters instantiating to themselves) and assuming all conditions.