Add.quick_simplified and quick_simplification implemented/tested
Add.quick_simplified was created to aid with ExprRange formatting.  The idea is that we can use "quick-n-dirty" tricks for formatting -- we don't have to prove the simplifications when it is just being used for window-dressing.  So this is fast and versatile (combining integers and making cancelations of terms, including portions of ranges of terms) without requiring explicit assumptions (we just assume terms represent numbers and expression ranges are well-formed).