Establish numbers/functions pkg and MonDecFuncs class
Related Issue: #97 (particularly related to the QPE theorem _fail_ineq). Establish numbers/functions package, including the common, demos, and theorems page, along with the MonDecFuncs class and a single related functions theorem. Creation of the package and creation of the MonDecFuncs class (insdie the mon_dec_funcs.py file) was prompted mostly by the need for a summation-to-integral bounding step in the QPE theorem _fail_ineq, which then needed the notion of a monotonically-decreasing function to be elaborated a bit.