Update Sum.deduce_in_number_set() to elim a preserve_all=True arg
Update Sum.deduce_in_number_set() to eliminate a preserve_all=True arg which was keeping some important simplification from happening. This was a remnant of code left over from when this was a standard pattern used when first deriving an implication (i.e. through the instantiation of an implication theorem) and then immediately deriving its consequent. The auto simplification process now handles this process correctly.