Add comment for future side-effect/automation augmentation
Add comment for future side-effect/automation augmentation: we want to augment the LessEq (or better, the NumberOrderingRelation class) with a method such as derive_one_side_in_real_subset() for when we have an ordering relation in which one side is equal to some literal number. This should help some automation.