Add Integrate.deduce_in_number_set() method
Add Integrate.deduce_in_number_set() method and comment out section of code in the __init__() method that was interpreting general number sets (such as Real, RealPos, etc) and converting them to Interval domains. The conversion of real number subsets to Interval domains was originally implemented to allow formatting of the integrals using lower and upper bounds in the graphical presentation but ended up being problematic for inferring closure properties (because we hadn't anywhere actually equated Reals, for example, with the (-infty, infty) interval). This issue should be followed up in the future with a formatting-based approach instead of literally replacing the integral's domain specification. An additional problem to be addressed: something like RealPos as a domain cannot simply be represented in the graphical depiction with the lower and upper boundaries of 0 to infty since the 0 is not included (such an integral would be more properly expressed as a limit from a to infty as a approaches 0 from the right).