New Add.simplification default to move irreducible values to the end.
Made an 'order_key' simplification directive for Add so it is easy to choose any canonical ordering for Add simplification.  The default just puts irreducible values at the end.  This implements half of Issue 215 (the other half is to implement this feature for Mult which requires a more significant change to its shallow_simplification routine).